requirements.txt
discord.py
asyncio
datetime
python_dotenv
quart
mcstatus
quart_discord
psutil
aiohttp